requirements.txt
langchain-core
langchain-community
langchain-perplexity
langserve

python-dotenv

streamlit
fastapi
uvicorn
sse_starlette